OJ-3158: Added If the disabled scenario
An error occurs when keyRotation is has been enabled but a disabled
flag is supplied. It tries to use original approach to decrypt the payload
but the payload was encrypted with un-aliased key and would through an excepted
invalid InvalidCiphertextException exception. The payload can not be decrypted
and validated because of the error CRI