OJ-3210: Categorise KMSRSADecrypterTest
Categorise the KMSRSADecrypterTest better highlight areas of testing concern
an extra test case for when key rotation is enabled and legacy flag is enabled
where initial description with an alias fails to decrypt, but the second alias
is used to decrypt successfully. Then it is expected the legacy decription does
not happen even though the legacy flag is set to true.