OJ-3210: Clarified step with flag enabled
Now that there are two flags, the step needs to be modified

Test should be updated such that when the key rotation flag is set to true i.e,
enabled and the legacy flag is set to false, It is clear the rotated key is used to
decrypt the request

Conversely, when the key rotation flag is set to false i.e disabled, it really does not matter
that the legacy flag is true or false, setting the key rotation flag would imply
using the legacy key to decrypt the request.

Reduce cyclomatic complexity of KMSRSADecrypter

SonarQube has flagged the KMSRSADecrypter class for high cyclomatic complexity.
This commit refactors the class to reduce its complexity by breaking down
the `decrypt`